+++ fiboa_cli/datasets/de_sax.py
@@ -21,9 +21,9 @@ PROVIDERS = [
 ]
 ATTRIBUTION = "Sächsisches Landesamt für Umwelt, Landwirtschaft und Geologie"
 LICENSE = "dl-de/by-2-0"
-EXTENSIONS = [
+EXTENSIONS = {
     "https://fiboa.github.io/flik-extension/v0.1.0/schema.yaml"
-]
+}
 COLUMNS = {
     'geometry': 'geometry',
     'FB_FLIK': ['id', 'flik'], # make flik id a dedicated column to align with NRW etc.