+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Rewrite to class-based converters
 - Use only unix line-endings in source files
+- Use set instead of list for Converter.extensions
 
 ## [v0.9.0] - 2025-01-07